Move process dependency to node-util.ts
