+++ examples/statemachine/package.json
@@ -64,7 +64,7 @@
         "vscode-languageserver": "~9.0.1"
     },
     "devDependencies": {
-        "langium-cli": "~3.0.2"
+        "langium-cli": "~3.0.3"
     },
     "volta": {
         "node": "18.19.1",
