+++ packages/langium-cli/src/generate.ts
@@ -17,8 +17,8 @@ import { generateModule } from './generator/module-generator.js';
 import { generateTextMate } from './generator/highlighting/textmate-generator.js';
 import { generateMonarch } from './generator/highlighting/monarch-generator.js';
 import { generatePrismHighlighting } from './generator/highlighting/prism-generator.js';
-import { elapsedTime, getTime, log } from './generator/langium-util.js';
-import { getUserChoice, schema } from './generator/node-util.js';
+import { getTime, log } from './generator/langium-util.js';
+import { elapsedTime, getUserChoice, schema } from './generator/node-util.js';
 import { RelativePath } from './package-types.js';
 import { getFilePath } from './package.js';
 import { validateParser } from './parser-validation.js';
