+++ packages/langium/package.json
@@ -70,7 +70,7 @@
     "vscode-uri": "~3.0.8"
   },
   "devDependencies": {
-    "langium-cli": "~3.0.2"
+    "langium-cli": "~3.0.3"
   },
   "volta": {
     "node": "18.19.1",
